gxm: Move code to renderer
