+++ src/emulator/CMakeLists.txt
@@ -43,6 +43,7 @@ add_subdirectory(mem)
 add_subdirectory(module)
 add_subdirectory(modules)
 add_subdirectory(nids)
+add_subdirectory(renderer)
 add_subdirectory(rtc)
 add_subdirectory(util)
 add_subdirectory(threads)
