+++ src/emulator/gxm/CMakeLists.txt
@@ -13,5 +13,5 @@ add_library(
 )
 
 target_include_directories(gxm PUBLIC include)
-target_link_libraries(gxm PUBLIC crypto glutil mem sdl2 rpcs3 threads vita-headers)
-target_link_libraries(gxm PRIVATE microprofile)
+target_link_libraries(gxm PUBLIC mem rpcs3 threads vita-headers)
+target_link_libraries(gxm PRIVATE)
