+++ src/emulator/gxm/include/gxm/texture_cache_state.h
@@ -1,24 +1,24 @@
 #pragma once
 
-#include <glutil/object_array.h>
-
-#include <array>
-
-struct SceGxmTexture;
-
-constexpr size_t TextureCacheSize = 1024;
-typedef uint64_t TextureCacheTimestamp;
-typedef uint32_t TextureCacheHash;
-
-typedef std::array<SceGxmTexture, TextureCacheSize> TextureCacheGxmTextures;
-typedef std::array<TextureCacheTimestamp, TextureCacheSize> TextureCacheTimestamps;
-typedef std::array<TextureCacheHash, TextureCacheSize> TextureCacheHashes;
-
-struct TextureCacheState {
-    size_t used = 0;
-    TextureCacheTimestamp timestamp = 1;
-    TextureCacheGxmTextures gxm_textures;
-    TextureCacheTimestamps timestamps;
-    TextureCacheHashes hashes;
-    GLObjectArray<TextureCacheSize> textures;
-};
+//#include <glutil/object_array.h>
+//
+//#include <array>
+//
+//struct SceGxmTexture;
+//
+//constexpr size_t TextureCacheSize = 1024;
+//typedef uint64_t TextureCacheTimestamp;
+//typedef uint32_t TextureCacheHash;
+//
+//typedef std::array<SceGxmTexture, TextureCacheSize> TextureCacheGxmTextures;
+//typedef std::array<TextureCacheTimestamp, TextureCacheSize> TextureCacheTimestamps;
+//typedef std::array<TextureCacheHash, TextureCacheSize> TextureCacheHashes;
+//
+//struct TextureCacheState {
+//    size_t used = 0;
+//    TextureCacheTimestamp timestamp = 1;
+//    TextureCacheGxmTextures gxm_textures;
+//    TextureCacheTimestamps timestamps;
+//    TextureCacheHashes hashes;
+//    GLObjectArray<TextureCacheSize> textures;
+//};
