+++ src/emulator/gxm/src/profile.h
@@ -1,5 +1,5 @@
 #pragma once
 
-#include <microprofile.h>
-
-#define GXM_PROFILE(name) MICROPROFILE_SCOPEI("GXM", name, MP_BLUE)
+//#include <microprofile.h>
+//
+//#define GXM_PROFILE(name) MICROPROFILE_SCOPEI("GXM", name, MP_BLUE)
