+++ src/emulator/host/CMakeLists.txt
@@ -19,5 +19,5 @@ version.cpp
 configure_file(src/version.cpp.in version.cpp)
 
 target_include_directories(host PUBLIC include)
-target_link_libraries(host PUBLIC audio cpu ctrl gxm io kernel mem net rtc util gui)
+target_link_libraries(host PUBLIC audio cpu ctrl gxm io kernel mem net renderer rtc util gui)
 target_link_libraries(host PRIVATE glbinding-aux glutil microprofile nids sdl2)
