+++ src/emulator/renderer/include/renderer/state.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace renderer {
+    struct State {
+    };
+}
