+++ src/emulator/renderer/include/renderer/texture_cache_functions.h
@@ -2,7 +2,7 @@
 
 //struct MemState;
 //struct SceGxmTexture;
-//struct TextureCacheState;
-//
-//bool init(TextureCacheState &cache);
+struct TextureCacheState;
+
+bool init(TextureCacheState &cache);
 //void cache_and_bind_texture(TextureCacheState &cache, const SceGxmTexture &gxm_texture, const MemState &mem, bool enabled);
