+++ src/emulator/renderer/src/profile.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <microprofile.h>
+
+#define R_PROFILE(name) MICROPROFILE_SCOPEI("renderer", name, MP_BLUE)
