+++ src/emulator/renderer/src/texture_formats.cpp
@@ -1,4 +1,4 @@
-#include <gxm/functions.h>
+#include <renderer/functions.h>
 
 //#include <util/log.h>
 //
